Move hook directory to src, Update docstrings, remove useless line
