+++ frontend/src/hooks/UseUser.ts
@@ -17,20 +17,10 @@ interface useUserReturn {
  * when one is logged in and accessing the user object appropriately if one is logged in.
  * Simply check if userIsLoggedIn and do work from there.
  *
- * userIsLoggedIn checks if a user is logged in. If one is not, it redirects the
- * user to the login screen.
- *
- * refreshUser re-fetches the user from the backend and redirects to the login screen
- * if one is not logged in.
- *
- * setUserToLoggedIn is a function that sets the state of the user to logged in
- *
- * setUserToLoggedOut is a function that sets the state of the user to logged out.
- *
  * @returns { user, userIsLoggedIn, refreshUser, setUserToLoggedIn, setUserToLoggedOut }
  */
 export const useUser = (): useUserReturn => {
-	const queryUser = (): TUser => {
+	const userQuery = (): TUser => {
 		// This is my current understanding of how to get a user object.
 		// I suspect I will be adding a get user endpoint to the backend soon.
 		const { data } = api.auth.getMe.useQuery();
@@ -41,25 +31,46 @@ export const useUser = (): useUserReturn => {
 				to: "/auth",
 			});
 		}
+
 		return data;
 	};
 
-	const [user, setUser] = useState<TUser>(queryUser());
+	const [user, setUser] = useState<TUser>(userQuery());
 	const [isLoggedIn, setIsLoggedIn] = useState(false);
 
-	const refreshUser = () => {
-		if (isLoggedIn) return;
-		setUser(queryUser());
-	};
-
+	/**
+	 * Gets an instance of the currently logged in user
+	 * If no user is logged in, it redirects to the login page
+	 *
+	 * @returns an instance of the currently logged in user
+	 */
 	const getUser = () => {
 		if (!isLoggedIn) throw redirect({ to: "/auth" });
 		return user;
 	};
 
+	/**
+	 * Refreshes the login for the current user
+	 */
+	const refreshUser = () => {
+		if (isLoggedIn) return;
+		setUser(userQuery());
+	};
+
+	/**
+	 * Sets logged in state to true
+	 */
 	const setUserToLoggedIn = () => setIsLoggedIn(true);
+
+	/**
+	 * Sets logged in state to false
+	 */
 	const setUserToLoggedOut = () => setIsLoggedIn(false);
 
+	/**
+	 * Checks if a user is logged in. If one is not, it redirects the
+	 * user to the login screen.
+	 */
 	const userIsLoggedIn = (): boolean => {
 		if (isLoggedIn) return true;
 
@@ -68,7 +79,6 @@ export const useUser = (): useUserReturn => {
 		});
 	};
 
-	setUser(queryUser());
 	return {
 		getUser,
 		userIsLoggedIn,
